Create main
Start of main function. 

Currently working on displaying boxplots in sequence, to be able to generate a control group (so as to be able to contrast results after the project). 

Moreover, boxplots are a well-established function within matplotlib, and serve as a good wireframe through which later operations can gain a foothold.